+++ static/js/core/mtu/_util/type-check.js
@@ -1,0 +1,11 @@
+export function isObject(obj) {
+  return Object.prototype.toString.call(obj) === "[object Object]";
+}
+
+export function isArray(array) {
+  return Array.isArray(array);
+}
+
+export function isElement(element) {
+  return element instanceof Element;
+}