+++ static/js/pages/st/study/study-course-container.js
@@ -1,6 +1,7 @@
 import { mtoEvents } from "../../../core/utils/mto-events";
+import { MtuSidebar } from "../../../core/mtu/sidebar/mtu-sidebar";
 import { mtuMenu } from "../../../core/ui/menu/mtu-menu";
-import { ApiCp } from "../../../core/api/api-cp";
+import { StCourseTree } from "../st-course-tree";
 /**
  *
  * Refactor mtmStudyContainer
@@ -9,38 +10,80 @@ import { ApiCp } from "../../../core/api/api-cp";
 export class StudyCourseContainer {
   constructor(options = {}) {
     this.options = options;
+    console.log(this.options);
     this.elThis = null;
 
-    this.selected_course_index = -1;
-
-    this.api = new ApiCp("cp", "course_n");
     this._init();
   }
-  _init() {
+
+  async _init() {
     this.elThis = document.createElement("div");
+
     this._initVariables();
 
-    this.clMenu = new mtuMenu();
+    const courseId = this.courseId;
+    const studentId = this.studentId;
+    const { courseBookData, courseData, resultData } = await this.getCourseData(courseId, studentId);
+    this.courseData = courseData;
+    this.resultData = resultData;
+    this.courseBookData = courseBookData;
 
-    const courseId = this.options.courseId;
-    this.getCourses(courseId);
+    this.initCourseTree({ courseBookData, courseData, resultData });
+
+    this.initSidebar();
   }
+
   _initVariables() {
     this.courseId = this.options.courseId ?? null;
     this.isDemo = this.options.demo ?? null; // annonymous?
     this.userType = this.options.userId ?? null;
+    this.studentId = this.options.studentId ?? null;
+  }
+
+  initSidebar() {
+    const learnTarget = this.elCourseTree;
+
+    this.clSidebar = new MtuSidebar({
+      position: "right",
+      items: [
+        { title: "학습하기", icon: "form", aside: learnTarget },
+        { title: "통계", icon: "barChart" },
+      ],
+    });
+    this.elSidebar = this.clSidebar.getElement();
+    this.options.rootElement.prepend(this.elSidebar);
+    // document.body.appendChild(this.elSidebar);
+  }
+
+  initCourseTree({ courseBookData, courseData, resultData }) {
+    const title = "학습하기";
+    this.clCourseTree = new StCourseTree({
+      courseBookData,
+      courseData,
+      resultData,
+      title,
+
+      onBranchClick: this.handleContentClick.bind(this),
+    });
+    this.elCourseTree = this.clCourseTree.getElement();
+    // this.elThis.appendChild(this.elCourseTree);
+    this.options.rootElement.appendChild(this.elCourseTree);
   }
   ////////////////////////////////////////////////////////////////////////////////////////
   ////////////////////////////////////// Handler  ////////////////////////////////////////
   ////////////////////////////////////////////////////////////////////////////////////////
   // Click Branch
-  handleContentClick(item) {
-    const contentData = item.data;
+  handleContentClick(data) {
+    const contentData = data;
     const param = {
+      student_id: this.studentId,
       course_id: this.courseId,
       content_id: contentData.id,
-      content_type: contentData.type,
       title: contentData.title,
+
+      content_type: contentData.type,
+      results: contentData.results,
+      units: contentData.units,
     };
     mtoEvents.emit("OnChangeCourseContent", param);
     // branchData.id;
@@ -48,96 +91,83 @@ export class StudyCourseContainer {
   ////////////////////////////////////////////////////////////////////////////////////////
   //////////////////////////////////////// URL  //////////////////////////////////////////
   ////////////////////////////////////////////////////////////////////////////////////////
-  async urlGetCourseTree(courseId) {
+  urlGetCourseBook(courseId) {
     try {
-      console.log(courseId);
-      const formData = new FormData();
-      formData.append("id", courseId);
-      return await axios.post("../cp/get-full-course/", formData).then((res) => {
-        console.log(res);
-        if (res.data.result) {
-          return res.data.result;
-        } else {
-          console.log("no data: ", res);
-          return [];
-        }
+      const url = `../cp/api/course_book/${courseId}`;
+      return axios.get(url).then((res) => {
+        return res.data ? res.data : {};
       });
-    } catch (err) {
-      console.log(err);
+    } catch (error) {
+      throw new Error(error);
     }
   }
-  async urlGetCourse(id, field_type) {
+  urlGetCourse(id, field_type) {
     try {
       const validFieldType = ["lists", "contents", "keys"];
-      if (validFieldType.indexOf(field_type) === -1) {
+      if (field_type && validFieldType.indexOf(field_type) === -1) {
         throw new Error(`Invalid field_type: ${field_type}`);
       }
 
       const param = {
-        id: id,
-        field_type: "lists",
+        field_type: field_type,
       };
-      const result = this.api.filter(param);
-      return result;
+      return axios
+        .get(`../cp/api/course_n/${id}/get_json_field/`, {
+          params: param,
+        })
+        .then((res) => {
+          if (res.data) {
+            return res.data;
+          }
+        })
+        .catch((err) => console.error(err));
     } catch (err) {
       console.log(err);
+    }
+  }
+
+  urlGetStudyResult(courseId, userId) {
+    try {
+      const param = {
+        id_student: userId,
+        id_course: courseId,
+        id_content: courseId,
+      };
+      console.log(param);
+      return axios.get("../st/api/study_result/", param).then((res) => {
+        const result = res?.data ? res.data : [];
+        return result;
+      });
+    } catch (err) {
+      throw new Error(err);
     }
   }
   ////////////////////////////////////////////////////////////////////////////////////////
   //////////////////////////////////////// API  //////////////////////////////////////////
   ////////////////////////////////////////////////////////////////////////////////////////
-  async getCourses(courseId) {
+  async getCourseData(courseId, studentId) {
     try {
-      const data = await this.urlGetCourseTree(courseId);
-      console.log(data);
-
-      this.courseId = data.id;
-      this.courseData = data;
-
-      // this.courseDropdown.setData(this.courseData);
-      // 아니면, this.mtuMenu.setData(this.courseData)
-      const menuItems = this.composeCourseItems(data);
-      this.clMenu.render(menuItems);
-      this.elThis.appendChild(this.clMenu.getRootElement());
-
-      // course list data (mCourse)
-      // this.courseData = await this.urlGetCourse(courseId, "lists");
-      // console.log(this.courseData);
-    } catch (error) {
-      console.log(error);
+      const courseBookData = await this.urlGetCourseBook(courseId);
+      const courseData = await this.urlGetCourse(courseId);
+      const resultData = await this.urlGetStudyResult(courseId, studentId);
+      return { courseBookData, courseData, resultData };
+    } catch (err) {
+      throw new Error(err);
     }
   }
-  composeCourseItems(data) {
-    const items = [];
-    const chapters = data.children;
-    const depth = 0;
-    chapters.forEach((chapter) => {
-      this.composeItem(chapter, items, "default", depth);
-    });
-    return items;
-  }
-  composeItem(data, parent, type, depth) {
-    const item = {
-      title: data.title,
-      type: type ?? "default",
-      data: data,
-    };
 
-    // branch
-    if (depth === 1) {
-      item.onClick = this.handleContentClick.bind(this);
-    }
-
-    if (data.children) {
-      const children = [];
-      const size = data.children.length;
-      const newDepth = depth + 1;
-      for (var i = 0; i < size; i++) {
-        this.composeItem(data.children[i], children, type, newDepth);
+  composeTreeData(lists) {
+    let treeData = [];
+    lists.forEach((data) => {
+      const type = data.type === 0 ? "chapter" : "branch";
+      if (type === "chapter") {
+        data["children"] = [];
+        treeData.push(data);
+      } else if (type === "branch") {
+        const chapterSize = treeData.length - 1;
+        treeData[chapterSize].children.push(data);
       }
-      item.children = children;
-    }
-
-    parent.push(item);
+    });
+    return treeData;
   }
 }